"Update client endpoint, project ID, and keys to use VITE_ENV_VARS"
This commit updates the initialization of the Appwrite client in the main.js file. It replaces the hardcoded endpoint, project ID, and keys with environment variables prefixed with "VITE\_". The new variables are: VITE\_API\_ENDPOINT, VITE\_API\_PROJECT\_ID, VITE\_API\_DATABASE\_KEY, and VITE\_API\_COLLECTION\_KEY. The console.log statements are added to print the values of these new variables.