Change hardcoded database and collection keys to dynamic variables; update API endpoint and project ID
The code diff above replaces the hardcoded `databaseId` and `collectionId` with dynamic variables `databaseKey` and `collectionKey`, respectively. It also updates the API endpoint and project ID to dynamic variables `apiEndpoint` and `projectId`. This change allows for more flexibility and customization in using different database and collection keys, as well as API endpoints and project IDs.